chore(frontend): moves state logic from Header.tsx to MainPage.tsx
